requirements.txt
streamlit==1.31.1
pickle-mixin==1.0.2
joblib==1.3.2
pandas==2.2.1
pillow==10.2.0
pip==24.0
streamlit-option-menu==0.3.12 
scikit-learn==1.2.2
xgboost==2.0.3
